requirements.txt
# Python Dependencies for SQL Server ML Project

# Database connectivity
pyodbc>=4.0.35
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Data quality and profiling
ydata-profiling>=4.6.0

# Optional: Advanced ML libraries
# xgboost>=1.7.0
# lightgbm>=4.0.0
# tensorflow>=2.13.0
# torch>=2.0.0
